archive/requirements.txt
pandas>=2.0.0
geopandas>=1.1.0
numpy>=1.24.0
soilgrids>=0.1.3
scikit-learn>=1.3.0
rasterio>=1.4.2
requests>=2.28.0
python-dotenv>=1.0.0
tenacity>=9.0.0
earthengine-api>=0.1.370
